2025-08-03 이민호 자재출고, 제품입고

+++ kbs/backend/src/main/java/com/kimbap/kbs/production/service/BomDetailVO.java
@@ -1,5 +1,7 @@
 package com.kimbap.kbs.production.service;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class BomDetailVO {
   private String bomDCd;      // BOM상세코드
-  private Integer needQty;    // 소요수량
+  private BigDecimal needQty;    // 소요수량
   private String unit;        // 단위
   private String isUsed;      // 사용여부
   private String mcode;       // 자재마스터코드
